Title: 숫자로 이루어진 사각형, Time: 112ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240831/숫자로 이루어진 사각형/rectangle-with-a-number.kt
@@ -10,7 +10,7 @@ class Main {
     private var num = 1
     
     fun solve() {
-        n = br.readLine()
+        n = br.readLine().toInt()
 
         repeat(n) {
             repeat(n) {
